feat: 앱 구조 변경
